refactor: move inline implementations from headers to .cpp files
Move non-template function implementations out of headers into
corresponding .cpp files to reduce header bloat, eliminate multiple
definition risks, and prepare for Arduino v3 compatibility.

Functions moved to src/types.cpp:
- str_sprintf (from globals.h)
- ULONGFromMemory, DWORDFromMemory, WORDFromMemory (from globals.h)
- formatSize (from globals.h) — also removes <sstream>/<iomanip> from
  the god header
- PreferPSRAMAlloc (from types.h)

Data moved to src/colordata.cpp:
- Color palette definitions (BlueColors_p, RedColors_p, GreenColors_p,
  RGBColors_p, spectrumBasicColors, spectrumAltColors, USAColors_p,
  rainbowPalette) moved from effectsupport.h with extern declarations
  in colordata.h

USE_M5 isolation:
- M5 audio logic extracted to m5audio.h/m5audio.cpp
- M5Screen extracted to screens/m5screen.h
- Reduces conditional compilation in globals.h

Builds verified locally: demo, m5stackdemo, mesmerizer
