+++ include/effectsupport.h
@@ -41,155 +41,8 @@
 #include <utility>
 #include "hashing.h"
 
-// Palettes used by a number of effects
+// Palettes used by a number of effects - now in colordata.h / colordata.cpp
 
-const CRGBPalette16 BlueColors_p =
-{
-    CRGB::DarkBlue,
-    CRGB::MediumBlue,
-    CRGB::Blue,
-    CRGB::MediumBlue,
-    CRGB::DarkBlue,
-    CRGB::MediumBlue,
-    CRGB::Blue,
-    CRGB::MediumBlue,
-    CRGB::DarkBlue,
-    CRGB::MediumBlue,
-    CRGB::Blue,
-    CRGB::MediumBlue,
-    CRGB::DarkBlue,
-    CRGB::MediumBlue,
-    CRGB::Blue,
-    CRGB::MediumBlue
-};
-
-const CRGBPalette16 RedColors_p =
-{
-    CRGB::Red,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-
-    CRGB::Red,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-
-    CRGB::Red,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-
-    CRGB::Red,
-    CRGB::DarkRed,
-    CRGB::DarkRed,
-    CRGB::OrangeRed
-};
-
-const CRGBPalette16 GreenColors_p =
-{
-    CRGB::Green,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-
-    CRGB::Green,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-
-    CRGB::Green,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-
-    CRGB::Green,
-    CRGB::DarkGreen,
-    CRGB::DarkGreen,
-    CRGB::LimeGreen
-};
-
-const CRGBPalette16 RGBColors_p =
-{
-    CRGB::Red,
-    CRGB::Green,
-    CRGB::Blue,
-    CRGB::Red,
-    CRGB::Green,
-    CRGB::Blue,
-    CRGB::Red,
-    CRGB::Green,
-    CRGB::Blue,
-    CRGB::Red,
-    CRGB::Green,
-    CRGB::Blue,
-    CRGB::Red,
-    CRGB::Green,
-    CRGB::Blue,
-    CRGB::Blue
-};
-
-const CRGBPalette16 spectrumBasicColors =
-{
-    CRGB(0xFD0E35), // Red
-    CRGB(0xFF8833), // Orange
-    CRGB(0xFFEB00), // Middle Yellow
-    CRGB(0xAFE313), // Inchworm
-    CRGB(0x3AA655), // Green
-    CRGB(0x8DD9CC), // Middle Blue Green
-    CRGB(0x0066FF), // Blue III
-    CRGB(0xDB91EF), // Lilac
-    CRGB(0xFD0E35), // Red
-    CRGB(0xFF8833), // Orange
-    CRGB(0xFFEB00), // Middle Yellow
-    CRGB(0xAFE313), // Inchworm
-    CRGB(0x3AA655), // Green
-    CRGB(0x8DD9CC), // Middle Blue Green
-    CRGB(0x0066FF), // Blue III
-    CRGB(0xDB91EF)  // Lilac
-};
-
-const CRGBPalette16 spectrumAltColors =
-{
-    CRGB::Red,
-    CRGB::OrangeRed,
-    CRGB::Orange,
-    CRGB::Green,
-    CRGB::ForestGreen,
-    CRGB::Cyan,
-    CRGB::Blue,
-    CRGB::Indigo,
-    CRGB::Red,
-    CRGB::OrangeRed,
-    CRGB::Orange,
-    CRGB::Green,
-    CRGB::ForestGreen,
-    CRGB::Cyan,
-    CRGB::Blue,
-    CRGB::Indigo,
-};
-
-const CRGBPalette16 USAColors_p =
-{
-    CRGB::Blue,
-    CRGB::Blue,
-    CRGB::Blue,
-    CRGB::Blue,
-    CRGB::Blue,
-    CRGB::Red,
-    CRGB::White,
-    CRGB::Red,
-    CRGB::White,
-    CRGB::Red,
-    CRGB::White,
-    CRGB::Red,
-    CRGB::White,
-    CRGB::Red,
-    CRGB::White,
-    CRGB::Red,
-};
-
-const CRGBPalette16 rainbowPalette(RainbowColors_p);
 
 // A pointer to the global effect factories.
 extern DRAM_ATTR std::unique_ptr<EffectFactories> g_ptrEffectFactories;
