+++ include/screens/m5screen.h
@@ -1,0 +1,59 @@
+//+--------------------------------------------------------------------------
+//
+// NightDriverStrip - (c) 2026 Plummer's Software LLC.  All Rights Reserved.
+//
+// This file is part of the NightDriver software project.
+//
+//    NightDriver is free software: you can redistribute it and/or modify
+//    it under the terms of the GNU General Public License as published by
+//    the Free Software Foundation, either version 3 of the License, or
+//    (at your option) any later version.
+//
+//    NightDriver is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//    GNU General Public License for more details.
+//
+//    You should have received a copy of the GNU General Public License
+//    along with Nightdriver.  It is normally found in copying.txt
+//    If not, see <https://www.gnu.org/licenses/>.
+//
+//---------------------------------------------------------------------------
+
+#pragma once
+
+#include "globals.h"
+#include "screen.h"
+
+#if USE_M5DISPLAY
+
+#include <M5UnitLCD.h>
+// M5Screen
+//
+// Screen class that supports the M5 devices
+
+class M5Screen : public Screen
+{
+  public:
+
+    M5Screen(int w, int h) : Screen(w, h)
+    {
+        M5.Lcd.fillScreen(GREEN16);
+    }
+
+    virtual void drawPixel(int16_t x, int16_t y, uint16_t color) override
+    {
+        M5.Lcd.drawPixel(x, y, color);
+    }
+
+    virtual void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color) override
+    {
+        M5.Lcd.fillRect(x, y, w, h, color);
+    }
+
+    virtual void fillScreen(uint16_t color) override
+    {
+        M5.Lcd.fillScreen(color);
+    }
+};
+#endif
